Fix duplicate license files in msi packages
It appears when we added the workloads subset, the msi projects
contained a step to add a License.txt file to the packages it creates.
This resulted in a conflict as runtime will also add one.

To fix, we'll skip adding the extra License.txt file when the project is
'msi.csproj'.

Note: A better fix might be to give msi.csproj a different extension,
but unsure of the implications.

Fixes https://github.com/dotnet/runtime/issues/56516