Remove unstable ThreadPool events test (#56680)
The threadpool events being verified by this test aren't
guaranteed to occur given the test code. All of them are
the results of background timers and relative execution
speed of work items. Given the relatively low event coverage
of the existing test and my historical understanding that
specific events don't often regress, trying to stabilize this test
is probably more hassle than it is worth.