Remove the 'all' option and implement 'rebuild' in cmd per Anton's PR feedback
It also turns out that the rebuild cleanup was in the wrong place
in the build.sh script - it was done as part of generate_layout i.e.
after the native and managed test components have been built i.o.w.
using it ended up building the tests and deleting them afterwards.

Thanks

Tomas