Allow ConfigBinder to bind arrays to Singular elements
When trying to bind a collection of type `T`, ConfigBinder will now try
to bind to the current object instead of children objects when the
provided configuration doesn't contain an array.

This change allows elements that are not arrays to be bound to arrays.
For example, a type string[] can now be bound to "key" along with
"key:0".