Improve AppendLiteral for short literals
DefaultInterpolatedStringHandler.AppendLiteral is special, in that the 99.9% use case is it's called as part of string interpolation by compiler-generated code where the argument is always a literal.  That means when the method is inlined, the JIT can specialize the implementation based on knowing the literal's Length is a constant.  Which means we can provide specialized implementations for common lengths, resulting in faster and smaller code.  This does so for strings of length 1 and 2, as they're quite common in code bases examined, and it does so for both the default handler and also the handler used to format into a span.

The exact same code as before gets generated by the JIT for literals of length > 2.  The main downside here is if someone does manually call this CompilerServices method with a variable, the call site will end up bloated with the special cases.