Define dependencies for aspnetcore transport pkg
The AspNetCoreApp internal transport packages relies on a set of
projects already being built. This was working well before when
the transport package was built after src.proj run.

Adding dependencies via ProjectReferences to the transport package
so that those are always built in before.

In a subsequent change, the output from these P2Ps should be used
instead of relying on binplacing.