[metadata] Generic params in covariant return checking is ok
The issue is that `mono_type_is_reference` does not look at whether a generic
parameter has a constraint that would make it a reference type (there is a big
comment near `mono_type_is_reference` that making it do so would break generic
sharing).

The issue is if you have something like:

```

class BaseClass {}

public class C {
       public virtual BaseClass SomeCovariantFunction() => ...
}

public class G<T> : C where T : BaseClass {
       public override T SomeCovariantFunction() => ...
}
```

Here when setting up `G<T>` we do a check that the override of
SomeCovariantFunction that returns T is a valid, which boils down to checking
that `T` is ok for a covariant return from `BaseClass`.  This is mostly the
same as `BaseClass` is-assignable-from `T`, except we also need to check that
`T` is not a valuetype that implements interface `BaseClass`.  That check was
wrong.  The fix is to check for `is_valuetype`, not `!is_reference` on `T`.

Fixes https://github.com/dotnet/runtime/issues/49731