Refactor `EQ`/`NE` morphing (#57671)
* Add fgUpdateConstTreeValueNumber

A helper to update VNs of constant trees in morph.

* Extract EQ/NE morphing into its own method

And modernize the code.

Zero diffs, both on 32 and 64 bit as expected.

* Remove a dead optimization

It is trying to eliminate what looks like
a CSE pattern, but what guarantees the temp is
last use? And it gives up (rightly) for actual
CSEs.

No diffs for any of the SPMI collections.

* Fix $1

Update const's VN for the MOD => AND optimization.

* Fix $2

Update const's VN for the X + A == B => X == B - A optimization.

* Fix $3

For the COMMA-RELOP transform, it is not easy to preverse VNs
because the COMMA needs to have its VN updated to not include
relopOp2's exceptions. So do it only during global morph.

Also assert that COMMA doesn't have GTF_REVERSE_OPS applied to it.

* Actually, just delete the COMMA-RELOP transform

Turns out it is also dead.

No diffs.

* Fix $4

Preseve VN for the RELOP(RELOP, 0/1) transform.
The returned node computes the same value the
original relop did.

* Fix $5

The AND-SHIFT transform cannot easily regenerate VNs.
So only perform it during global morph.

Zero diffs.

* Fix $6

Bash the node and update its VN for the EQ(CAST(long <- int), CNS_LNG)
-> EQ(int, CNS_INT) transform.

* Fix $7

Use signed types to avoid potential issues with
missed sign extensions when cross-compiling.

* Stop propagating the relop type

It is no longer necessary as the TYP_BYTE setcc
optimization has been moved to lowering.

* Port fix from the main branch

This work is older than the aforementioned fix...

* Cleaning up - get rid of "cns2"

* Cleaning up - strongly-type trees

* Cleaning up - use matching types

Using ssize_t's instead of size_t's as that's
the type the node uses to store the value.

* Strongly type "op2" as a constant node

* Cleaning up - renames and shuffles

Also move some conditions and asserts around.

* Add a clarifying comment