Don't regress SkipLocalsInit optimization on net5.0 assets
net5.0 assets which are part of the Microsoft.NETCore.App shared framework and ship inside a package need to keep the SkipLocalsInit optimization to avoid regressing perf. This would happen when referencing a 6.0 package but consuming the net5.0 asset in it. The higher assembly version of the net5.0 asset would win over the same asset in the 5.0 shared framework and without this fix, the SkipLocalsInit optimization would be lost.

Contributes to https://github.com/dotnet/runtime/issues/54964