[mono] Don't generate safepoints inside write barrier wrapper (#58346)
* [mono] Skip m2n wrapper for mono_gc_wbarrier_generic_nostore_internal

There is no need for a m2n transition since the method just marks the card table. No exception can be thrown, no marshalling is required, no gc can happen during its execution etc. In addition to the redundancy of the wrapper, it also contains gc safepoints which is wrong. Having a GC between the ref store and the corresponding card table marking can in theory crash the GC.

* [mono] Disable safepoints in write barrier wrapper

This wrapper is called immediately after a store to a ref field. The ref store and the card table marking should be atomic in relation to GCs. Therefore we shouldn't have a safepoint in the wrapper.

* [mono] Refactor disabling of safepoints on GC wrappers