Delete `TruncateOrSignExtend32` (#56904)
* Introduce SetValueTruncating

As a less error-prone replacement of TruncateOrSignExtend32.
Add a comment header explaining its intended use case.

* Case $1

The code truncated, for unclear reasons (likely a copy & paste mistake),
the wrong node.

* Case $2

* Case $3

The old code was checking for the type of the shift node,
which is what we want. So that SetValueTruncating sees this
type, set it before setting the value itself.

* Delete TruncateOrSignExtend32

Constant integer nodes never have GTF_UNSIGNED set on them.
Additionally, zero-extending a TYP_INT const is incorrect.