handle the content-encoding can lie
Sometimes the server lies to us and says that the response is gzip compressed, but it is not. 

this then leads to an exception.

my suggestion is to check in advance if the content stream really starts with the signature `1F 8B 08` and if not, to treat it as a normal uncompressed stream.

the submission is to be understood as proof of concept and has not been tested yet !

### how to reproduce this:

create an http response with the following content. dont compress the content!

```
HTTP/1.1 200
content-encoding: gzip
Content-Type: text/html;charset=utf-8
Content-Length: 4

test
```