JIT: fix jump threading to keep track of preds instead of reanalyzing (#58579)
In #58368 we see an issue where a previously ambiguous pred becomes a true pred
once other pred flow is redirected. While this may be a missed opportunity, we
are better off just keeping with our initial assessment for now.

So during the initial pred walk, track which preds are ambiguous and true (and
by implication, false) and then use that classification during the second walk.

Fixes #58368.