Ensure proper memory ordering when background_ephemeral_sweep copies over the free list data that it found for gen 0. (#58564)
The method allocator::copy_with_no_repair is called at the end of background_ephemeral_sweep to copy over the free list data that it found for gen 0. There is no synchronization, but we are careful to copy items in a sequence that makes sure things work correctly for x86 and x64. However, on other platforms, we need a memory barrier.

This is a tentative fix for issue #57956 - it seems possible that the issue is caused by the lack of synchronization in copy_with_no_repair, but there is no proof yet.