Split and rework fgMorphCast
Split it into the part that does pre-order transforms
and the part that does post-order optimizations.

Rename the pre-order part to "fgMorphExpandCast" and let
"fgMorphSmpOp" do the morphing of operands and optimization.

Just one diff for this commit: "gtFoldExprConst" does not
retain the types of handles, thus LSRA was not seeing
"long" and "byref" constant integers with the same values as
equivalent and did not reuse the register.