Move Android SDK logic to /eng
Moves the Android SDK logic to a common location so that it can be used in the entire repo.

In particular, it enables targeting Android from the coreclr partition.

Mono could benefit from this move too, but it currently rolls its own copy of everything (not just Android).

The `export CLR_CC=$(which false)` part felt like a hack so I attempted to clean it up.