Expand ARM64 candidates for tailcall
Also add some assertions that we are using non-volatile registers when
generating the tailcall.

Fixes the arm64 tests as a side effect. It seems there is a bug in the
allocator when specifying only a single register as a candidate where it
may use a local variable's home instead of the candidate.