[System.IO.Hashing] Expose methods to get 32-bits and 64-bits hash values
I can see why the `GetCurrentHash` method returns `byte[]` (to be flexible with the size of the different hashes) but given the currently supported non cryptographic hashes (crc32, crc64, xxHash32 and xxHash64) it makes sense to return 32-bits and 64-bits value as respectively `int` and `long`. (`uint` and `ulong` are ruled out because they would not be CLS compliant)

This commit introduces two new intermediate abstract classes, `NonCryptographicHashAlgorithm32` and `NonCryptographicHashAlgorithm64` that expose respectively `int GetHash() ` and `long GetHash()`.

Crc32 and XxHash32 now inherit from `NonCryptographicHashAlgorithm32`.

Crc64 and XxHash64 now inherit from `NonCryptographicHashAlgorithm64`.

Maybe this design is not the most appropriate but in one way or another I think getting the hash as a 32-bit or 64-bit value is something that should definitely go into System.IO.Hashing.

As a consumer of this new System.IO.Hashing package I was *very* surprised that this was not available.