Treat ALWAYS of callf+always pair normally and fix live range coalescing
It is possible that the live-out set of the always block of a
callf+always pair is not a subset of the live-out of the callf block.
This can happen when the finally has an infinite loop, for example. In
that case we should not skip the always block in genCodeForBBlist as
then we will not open the correct variable ranges, which can lead to
wrong live var ranges and asserts.

To fix this change the logic so that we do not skip over the always
block. However, still skip code-gen for it as that is handled by
genCallFinally already. Note that this technically means that we emit
the live-range updates for the always *after* we have emitted the
instruction corresponding to it, but this is still less of a problem
since that block is going to be empty or a single jump/nop anyway that
does not affect liveness.

This introduced quite a few debug info changes since we are now opening
some ranges twice (one for the always and one for the block that
follows, which end up at the same emitter location). After looking at
the logic that coalesces live ranges for a while I noticed that it seems
wrong: the end location of live-ranges is exclusive, so
emitLocation::IsPreviousInsNum does not seem like the right check, as it
will coalesce two ranges that are separated by one instruction.

I also noticed that we do not properly coalesce two subsequent equal
live-ranges if the first one is the prolog. That's because the loop in
genSetScopeInfoUsingVariableRanges reports prolog ranges separately. To
fix the previous coalesce logic and improve the prolog case I have moved
all the logic into genSetScopeInfoUsingVariableRanges that now keeps
track of the previous range and makes sure to only report one range.

The improved coalescing is still enough to make up for the more accurate
live ranges for try-finally: the size of SPC is reduced by around 0.3%
with this change (11,549 KB -> 11,510 KB).

Fix #57752