Use less calls for MoveDirectory
See #58240
This fix is trying to overcome the concept of matching a universal behaviour inside a platform aware implementation.

All checks for consistent behaviours can be done within the public Directory implementation.
The missing bits which are only relevant for Unix are being left e.g. matching trailing seperators on both paths.

This might be redundant, given that Windows would handle such cases, so we dont have to do the checks beforehand, but given that we can reduce calls on both Unix and Windows runtimes should be a slight improvement.
There is only a single FileExist call for Unix and for windows we won't execute the move operation and wait for exceptions.