[wasm] Fix support for AOT with trimming disabled
When publishing with AOT (`RunAOTCompilation=true`), the assemblies get
trimmed by the linker, and copied to a single `linked/` folder. Then
`MonoAOTCompiler` runs `mono-aot-cross` on those assemblies. Since, all
the assemblies have their dependencies in the same folder too,
`mono-aot-cross` can resolve them from there.

But when trimming is disabled (`PublishTrimmed=false`),
`MonoAOTCompiler` gets paths to assemblies from different folders. For
example, some might be from `obj`, nugets etc. Which means that
`mono-aot-cross` might not always be able to find dependencies for the
assemblies being AOTed.

To fix this, if all the input assemblies are not in the same directory,
then we copy them to a temporary one, and then run `mono-aot-cross` on
them.

Fixes #53804