Change names of functions to improve readability and refactor code
This commit changes and refactors the core functions of this PR to have clearer names and to create a more logical flow of the code. It also updates some comments and fixes some oversights. Changes are based on the discussion https://github.com/dotnet/runtime/pull/49555#discussion_r705180391 (read whole conversation).