MonoAOTCompiler: check for nothing-changed case
The task needs to run `mono-aot-cross` on all the assemblies, even if a
given assembly is older than its output files, so as to correctly handle
changes in its dependencies.

This commit tries to check if all the input assemblies are up-to-date with their
corresponding output files, and avoids doing any work if that's true.