Process.Unix: handle ETXTBSY error when executing file that was just written by .NET.
ETXTBSY means we're trying to execute a file that is open for write.
This may be a file that was just written and closed, but is still referenced by another process
that has forked but not yet exec-ed.

When this error occurs, we wait for all processes to finish exec-ing, which causes the file
to be closed (CLOEXEC). Then we start the process again.