Fix BinSkim warnings in System.IO.Compression.Native.dll
Fixes "BA2007: 'System.IO.Compression.Native.dll' was compiled at too low a warning level (effective warning level 0 for one or more modules). Warning level 3 enables important static analysis in the compiler to flag bugs that can lead to memory corruption, information disclosure, or double-free vulnerabilities. To resolve this issue, compile at warning level 3 or higher by supplying /W3, /W4, or /Wall to the compiler, and resolve the warnings emitted.".

We used to have a `/W0` here that for whatever reason was acceptable for BinSkim in the past releases. That was deleted in #47344.

This BinSkim warning was reported in NativeAOT, but I assume this affect the 6.0 release as well.