[interp] Add stsfld.vt.w interpreter opcode
For methods that require more than 64k data items, add a variant stsfld
instruction that stores data item indices as 32-bit integers instead of 16-bit.

To solve this problem in general, we need wide variants of more opcodes, but
for specific (machine generated) methods, a wide stsfld instruction is sufficient.

Resolves https://devdiv.visualstudio.com/DevDiv/_workitems/edit/1398069