Work around for forbidden "Roslyn4.0" in test name on Android
Android apk names must be valid Java class names, and It is forbidden in
Java for a class name component to have a leading numeric digit. So
whilst "net.dot.foo3_1" is legal, "net.dot.foo3.1" is illegal. Use a
Regex to detect and rewrite all cases of this in AndroidAppBuilder