Only set path in x64 mac installer when installing on x64 (#59210)
* Only set path in x64 mac installer when installing on x64

In other words don't set the path for x64 installer on ARM64

* Remove postinstall script from hostfxr

There should be no need for both the host and hostfxr to set the path to dotnet.

Since the host installs dotnet, it should be the only package responsible for this.

* Make postinstall set install_location on mac

Also refactor script to use a template so that we don't need to fork the script.

* fix some syntax errors in script

* Update src/installer/pkg/sfx/installers/dotnet-host.proj

Co-authored-by: Adeel Mujahid <3840695+am11@users.noreply.github.com>

* Refine uname regular expressions

Co-authored-by: Adeel Mujahid <3840695+am11@users.noreply.github.com>